pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-cli"
version = "0.1.0"
description = "CLI agent for corporate AI chat"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tiktoken>=0.7",
    "jinja2>=3.1",
    "nbformat>=5.9",
    "pyjwt>=2.8",
    "prompt-toolkit>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "respx>=0.21",
    "types-PyYAML>=6.0",
]

[project.scripts]
genai = "genai_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
genai_cli = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=genai_cli --cov-report=term-missing"
